Added Utility resolvers - refactor
